controller: javascript variable

+++ controller/inc/class.uicase.inc.php
@@ -1022,7 +1022,7 @@
 
 		function add_case()
 		{
-			$set_completet_checklist = phpgw::get_var('set_completet_checklist', 'bool');
+			$set_completet_checklist = (int)phpgw::get_var('set_completet_checklist', 'bool');
 			$config = CreateObject('phpgwapi.config', 'controller');
 			$config->read();
 			$mandatory_location = $config->config_data['control_mandatory_location'];
@@ -1147,7 +1147,6 @@
 //			_debug_array($data);die();
 			phpgwapi_jquery::load_widget('core');
 
-			
 			$js = <<<JS
 				var enable_add_case = true;
 				var set_completet_checklist = {$set_completet_checklist};
